Fixed bug with hit testing.
Needed to override hitTestSelf() to return true, indicating that
a scroll pane is hit for events within its bounds even when a child
doesn't have a hit. Without this, objects such as RenderPadding,
which don't report hits in the padding region, would cause the
scroll pane to not receive event dispatches for trackpad events that
occured over those regions.