Fixed scroll bar layout
Updated it to use ScrollBarConstraints and the associated layout
model, as opposed to having the weird layoutWithValues() method.
The latter was causing the scroll bar to think that no layout
was necessary when the box constraints stayed the same even if
the values changed, which meant that if a scroll pane's view
changed size without the viewport changing size, the scroll bars
would not respond.