Improve `RenderScrollPane` layout
Rather than asking for the view's intrinsic dimensions, then laying
the view out again, we can just lay the view out in the first place.
This optimizes layout some, but it also ensures that our `viewWidth`
and `viewHeight` values match reality in the event that the view has
a bug in its intrinsics calculations.