Improvements to DeferredLayoutMixin
Whether to defer markNeedsLayout() depends on whether the action
the initiated the sequence of events leading up to the markNeedsLayout
is an input event that might be delivered in rapid succession (such
as hover or scroll events).  The actual markNeedsLayout() calls shouldn't
be concerned over whether to defer, but rather the entry points to those
events.

This changes the mixin to provide a method whereby callers can run a
callback inside a block that will catch all calls to markNeedsLayout()
and defer them.