Moved to broadcast streams
Single-subscription streams would yield errors when the tree
was updated using a globalKey, in which case a render object
can be detached and re-attached, causing us to try to listen
to a stream more than once. Such a case can happen easily
when in developer mode and enabling/disabling the performance
overlay.