Remove ListViewSelectionController.attach & detach
They served to enforce that the caller can't set the selection
to an out-of-bounds index based on an attached list view,
but that's notnecessary at all -- the caller just sets selected
ranges, and if those ranges are outside the bounds of the actual
list content, there's no inherent harm in that; it's the caller's
business.

The only non-assert usage of attach/detach was the `selectAll()`
API method, but that API's not needed either, since the caller
knows their data model and can just set the desired range.

Fixes #14